Move SharingShortcuts component to managingActions as it is only used there
