+++ app/packs/src/components/managingActions/SharingShortcuts.js
